Fixing extra README, adding factories, a bug.
Might as well remove the markdown README, it'll just get out of date.

Made a factory for a small User class. We'll need it for some tests, but
we don't want to give one away with extraextra.

read_all_about_it was just returning a hash rather than an Extra object,
so we made it do that.